LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/after-transition-reload-expected.txt


FAIL scroll: after-transition should work on a reload navigation assert_equals: expected 18 but got 431
